Allow range of seeds to be specified.
This allows the user to specify a range of seeds instead of always
checking randomised seeds.  Both Minimum and Maximum seeds are
specified, and must be numerical (long).

When a given range of seeds has been checked, as is with the Seeds to
Find setting, the run will stop.

This is also quite handy for Bedrock users, as it will allow them to
limit searches from 0 to 2147483648, as seeds within that rangs in both
Java and Bedrock are compatible.