Remove unnecessary space after shadow seed
I just kept on copying the space of the shadow seed on accident, which got annoying, since a space is 'hard to see'.
The space there is also unnecessary, so having it there doesn't make sense. Typing this took more time than actually making this pull request.  
I have not tested this code because I love living on the edge. But with my expertise I am quite certain it works as intended without testing.