Add two separate batch files - build and run for easier testing
Not strictly necessary but makes workflow a bit easier by allowing 
double click to build or run (on Windows). On top of #87. The file names 
of course don't have to be exactly this but I thought it was the most 
clear. Any changes to the exact titles are of course welcome.